spike/docker-python/pyproject.toml
[project]
name = "docker-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dotenv>=0.9.9",
    "pyyaml>=6.0.3",
]
